added new feature tests

+++ database/migrations/2021_04_01_191003_create_books_table.php
@@ -14,7 +14,7 @@ class CreateBooksTable extends Migration
     public function up()
     {
         Schema::create('books', function (Blueprint $table) {
-            $table->integer('isbn', 13)->autoIncrement();
+            $table->bigInteger('isbn', 13)->autoIncrement();
             $table->string('title', 100);
             $table->date('reserved_from');
             $table->date('reserved_to');
